+++ drivers/reset/reset-socfpga.c
@@ -11,12 +11,11 @@
 #include <linux/of_address.h>
 #include <linux/platform_device.h>
 #include <linux/reset-controller.h>
+#include <linux/reset/reset-simple.h>
 #include <linux/reset/socfpga.h>
 #include <linux/slab.h>
 #include <linux/spinlock.h>
 #include <linux/types.h>
-
-#include "reset-simple.h"
 
 #define SOCFPGA_NR_BANKS	8
