+++ drivers/reset/reset-sunxi.c
@@ -14,12 +14,11 @@
 #include <linux/of_address.h>
 #include <linux/platform_device.h>
 #include <linux/reset-controller.h>
+#include <linux/reset/reset-simple.h>
 #include <linux/reset/sunxi.h>
 #include <linux/slab.h>
 #include <linux/spinlock.h>
 #include <linux/types.h>
-
-#include "reset-simple.h"
 
 static int sunxi_reset_init(struct device_node *np)
 {
